pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TradingStrategist"
version = "0.1.0"
description = "Machine learning strategies for algorithmic trading"
readme = "README.md"
authors = [
    {name = "Wes", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial :: Investment"
]
requires-python = ">=3.8"
dependencies = [
    "numpy==1.24.3",
    "pandas==2.0.3",
    "matplotlib==3.7.2",
    "scikit-learn==1.3.0",
    "pyyaml==6.0.1",
    "scipy==1.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.0",
    "black==23.7.0",
    "flake8==6.1.0",
    "jupyter==1.0.0",
    "isort==5.12.0",
    "mypy==1.4.1",
]
docs = [
    "sphinx==7.1.1",
    "sphinx-rtd-theme==1.2.2",
]
interactive = [
    "jupyter==1.0.0",
    "ipywidgets==8.0.7",
]

# Command-line entry points
[project.scripts]
train-strategy = "TradingStrategist.train:main"
evaluate-strategy = "TradingStrategist.evaluate:main"

[project.urls]
"Homepage" = "https://github.com/yourusername/tradingstrategist"
"Bug Tracker" = "https://github.com/yourusername/tradingstrategist/issues"
"Documentation" = "https://tradingstrategist.readthedocs.io/"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*", "examples*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
filterwarnings = [
    "ignore::DeprecationWarning:sklearn.*:",
    "ignore::FutureWarning:pandas.*:"
]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false